requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0
django-daraja==1.3.0
django-decouple==2.1
django-import-export==3.3.4
django-jazzmin==2.6.0
et-xmlfile==1.1.0
idna==3.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
psycopg2==2.9.9
pycparser==2.21
python-decouple==3.8
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
urllib3==2.1.0
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
